refactor: consolidate changelog prompts
