--- apps/dashboard/src/lib/ai/prompts/changelog/conversational.ts
@@ -1,8 +1,0 @@
-import { buildChangelogPrompt } from "./base";
-import { toneConfigs, type ChangelogPromptParams } from "./types";
-
-export function getConversationalChangelogPrompt(
-  params: ChangelogPromptParams,
-): string {
-  return buildChangelogPrompt(params, toneConfigs.Conversational);
-}
