--- apps/dashboard/src/lib/ai/prompts/changelog/formal.ts
@@ -1,8 +1,0 @@
-import { buildChangelogPrompt } from "./base";
-import { toneConfigs, type ChangelogPromptParams } from "./types";
-
-export function getFormalChangelogPrompt(
-  params: ChangelogPromptParams,
-): string {
-  return buildChangelogPrompt(params, toneConfigs.Formal);
-}
